+++ HopperRender/resource.h
@@ -24,8 +24,9 @@
 #define IDC_BLURREDFRAMES 			    1025
 #define IDC_SCENECHANGETHRESHOLD        1026
 #define IDC_CURRFRAMEDIFF               1027
-#define IDC_SIDEBYSIDE1                 1028
-#define IDC_SIDEBYSIDE2                 1029
+#define IDC_CURRFRAMEDIFFDESC           1028
+#define IDC_SIDEBYSIDE1                 1029
+#define IDC_SIDEBYSIDE2                 1030
 
 // Next default values for new objects
 //